face-recognition/requirements.txt
flask==3.0.0
face-recognition==1.3.0
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
psycopg2-binary==2.9.9
requests==2.31.0

